+++ spykeviewer/api.py
@@ -28,6 +28,8 @@ class __ConfigOptions:
         self.autoselect_channels = True
         # Select all visible units in navigation by default
         self.autoselect_units = False
+        # Tranformation function for file paths when starting plugin remotely
+        self.remote_path_transform = lambda x: x
 
     def __setitem__(self, key, value):
         self.__dict__[key] = value
@@ -54,23 +56,26 @@ def start_plugin(name, current=None, selections=None):
         selections. If ``None``, the regular selections from the GUI
         are used.
     """
-    return window.start_plugin(name, current, selections)
+    return window.start_plugin(name, current, selections, False)
 
 
-def start_plugin_remote(name, current=None, selections=None):
+def start_plugin_remote(plugin, current=None, selections=None):
     """ Start first plugin with given name using the remote script.
     Raises a SpykeException if not exactly one plugins with
     this name exist.
 
-    :param str name: The name of the plugin. Should not include the
-        directory.
+    :param plugin: The name of the plugin or the plugin object. If this
+        is a string, it should not include the directory and the
+        plugin with this name will be started using the remote script.
+        If it is a plugin object, this plugin will be started using
+        the remote script.
     :param current: A DataProvider to use as current selection. If
         ``None``, the regular current selection from the GUI is used.
     :param list selections: A list of DataProvider objects to use as
         selections. If ``None``, the regular selections from the GUI
         are used.
     """
-    window.start_plugin_remote(name, current, selections)
+    window.start_plugin_remote(plugin, current, selections)
 
 
 def get_plugin(name):