+++ spykeviewer/ui/main_window_neo.py
@@ -18,11 +18,12 @@ from spykeutils.plugin.data_provider_neo import NeoDataProvider
 from spykeutils.plugin.data_provider_stored import NeoStoredProvider
 from spykeutils.plugin import io_plugin
 
-from main_window import MainWindow
+from .main_window import MainWindow
 from ..plugin_framework.data_provider_viewer import NeoViewerProvider
-from neo_navigation import NeoNavigationDock
-from dir_files_dialog import DirFilesDialog
-import io_settings
+from .neo_navigation import NeoNavigationDock
+from .dir_files_dialog import DirFilesDialog
+from . import io_settings
+from .. import api
 
 logger = logging.getLogger('spykeviewer')
 
@@ -561,6 +562,8 @@ class MainWindowNeo(MainWindow):
             sl.append(s.data_dict())
 
         io_plugin_files = []
+        transform_path = getattr(api.config, 'remote_path_transform',
+                                 lambda x: x)
         for s in sl:
             if s['type'] != 'Neo':
                 continue
@@ -572,6 +575,7 @@ class MainWindowNeo(MainWindow):
                                         '%s is not a known IO class!' % b[2])
                     if getattr(io, '_is_spyke_plugin', False):
                         io_plugin_files.append(io._python_file)
+                b[1] = transform_path(b[1])
 
         selections = json.dumps(sl, sort_keys=True, indent=2)
         config = pickle.dumps(plugin.get_parameters())